removed typescript from project
